feat: add scalar schema
- adds scalar docs
- adds test ci for codecov to see how it works (dont merge)
- renames function to less (?) confusing name for the dummies like me who thought it was a typo
- adds deprecation warning for ReDoc starting in 2.5.0

### TODO

The Scalar Litestar-custom css should be (?) loaded as a string from its home in `openapi/styles/scalar.css`..
do you rely on the StaticFilesConfig here?

Will have to see if we instantiate one even if a user doesnt pass the config into their Litestar app object.
If a user isnt using it, and we dont, should we generate one so that we can load this based on the static files url?